clean up the source code
